Allow bundle viz to work with dependencies satisfied by prereleases.
Suppose a Gemfile specifies a dependency, `foo`:

  foo

Let's say `foo` requires a certain `bar`:

  foo
    bar (>= 0.5)

If this dependency happens to be satisfied by a prerelease `bar`, e.g.,
0.6.pre, then `bundle viz` will error.

The cause of this error is that `Bundler::Graph#_populate_relations`
executes

  child_dependencies = dependency.to_spec.runtime_dependencies.to_set

and if `dependency` (here, `bar`), is a prerelease, then
`dependency.to_spec` will be `nil` on Rubygems 2.4.0 and higher because
of changes to `#to_spec`.

By forcing `dependency.prerelease = true`, then `to_spec` won't ignore
the prerelease `bar` dependency, which prevents `dependency.to_spec` to
from being `nil`.

This should address GitHub issues 3621 and 3217.